refactor: update structure of folders
