+++ components/features/Diary/AddNewNoteSection/AddNewNoteSection.tsx
